+++ CHANGELOG.md
@@ -40,6 +40,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - Local database stack and operations documentation.
 - Python classes and TypeScript interfaces for CRUD editor JSON config files validation and update the corresponding guide [GS-172].
 - How to create tables and forms documentation [GS-172].
+- Api Keys to User Profile on exampleapp and fastapitemplate [GS-251].
 
 ### Changed
 - Enhance "exampleapp/apps/mcp-server/run_mcp_server.sh" separating the SCRIPT_DIR and BASE_DIR envvars.
@@ -60,7 +61,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - MkDocs install scripts removes the .venv to include latest dependencies version.
 - Serve and run in main Makefile changed so "make run" makes a complete clean, regeneration and "make serve", and "make serve" only runs "mkdocs serve".
 - Upgrade Lambda runtime to Python 3.12, update API Gateway to OpenAPI 3.0.1 with CORS, and refactor endpoint definitions in "aws_big_lambda/template-sam.yml" [GS-245].
-- Move "parentUrl" attribute from "parentKeyNames" to the root of the JSON config files [GS-172].
+- Rename "parentKeyNames" to "endpointKeyNames" in JSON config files [GS-159].
+- Move "parentUrl" attribute from "endpointKeyNames" to the root of the JSON config files [GS-159].
 - Rename AI chatbot "GPT functions" to "AI Tools" in exampleapp comments.
 - Bind mkdocs serve to localhost:8015 to avoid conflicts with other GS APIs [GS-172].